app/pyproject.toml
[tool.poetry]
name = "budgly"
version = "0.1.0"
description = "Server API to track expenses and do some KPI"
authors = ["Alexis THIBAULT <alexis.thibault31@gmail.com>"]
license = "LICENSE"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = { extras = ["standard"], version = "^0.29.0" }
fastapi = "^0.111.0"
pydantic = "^2.7.1"
sqlmodel = { extras = ["asyncio"], version = "^0.0.18" }
asyncpg = "^0.29.0"
alembic = "^1.13.1"
pydantic-settings = "^2.2.1"
tenacity = "^8.2.3"
firebase-admin = "^6.5.0"
httpx = "^0.27.0"
google-auth-oauthlib = "^1.2.0"
logfire = "^0.30.0"
opentelemetry-instrumentation-asgi = "^0.45b0"
opentelemetry-instrumentation-fastapi = "^0.45b0"
pydantic-extra-types = "^2.7.0"
logging = "^0.4.9.6"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
mypy = "^1.9.0"
ruff = "^0.3.7"
coverage = "^7.4.4"
pre-commit = "^3.7.0"
pytest-cov = "^5.0.0"
sqlalchemy = { extras = ["mypy"], version = "^2.0.29" }
pytest-mock = "^3.14.0"
types-oauthlib = "^3.2.0.20240217"
pytest-xdist = "^3.6.1"
pytest-asyncio = "^0.23.6"
asgi-lifespan = "^2.1.0"
pyjwt = "^2.8.0"
polyfactory = "^2.16.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
cache_dir = ".pytest_cache"
python_files = "tests_*.py"
#addopts = "-n 4"
asyncio_mode = "auto"

[tool.coverage.run]
concurrency = ["greenlet"]
omit = [
    "alembic/*",
    "**/__init__.py",
    "scripts/*",
    "tests/*",
    "core/*"
]

[tool.mypy]
strict = true
exclude = ["alembic", "tests"]
plugins = [
    "pydantic.mypy",
    "sqlalchemy.ext.mypy.plugin"
]

[tool.ruff]
target-version = "py310"
exclude = ["alembic"]

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "B",
    "C4",
    "UP",
]
ignore = [
    "E501",
    "B008",
    "W191",
    "B904",
]

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true

[tool.logfire]
project_name = "budgly-api"
console = false
